refactor(examples): split small example responsibilities
